Use a compute shader to do image post processing asynchronously
The unpremultiplication and transparent pixel counting is done on the GPU in the command buffer right after the camera is rendered. By doing so, these processes are sometimes even ready before the next update loop. Transparent counts are requested asynchronously after the compute shader call is dispatched, and the textures and transparent counts are queued up to be retrieved one at a time in the update loop. This gets the time spent grabbing textures and creating sprites down to a mere 50 microseconds.

It appears that the render setup process takes a pretty hefty portion of the frametime now.